requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
ddddocr==1.4.7
Flask==2.1.2
flatbuffers==23.5.9
fonttools==4.33.3
greenlet==2.0.2
humanfriendly==10.0
idna==3.4
importlib-metadata==4.11.4
importlib-resources==5.12.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.7.1
mpmath==1.3.0
mysql==0.0.3
mysql-connector-pyhton==0.1
mysql-connector-python==8.0.33
mysqlclient==2.1.1
numpy==1.24.3
onnxruntime==1.14.1
opencv-python-headless==4.7.0.72
packaging==21.3
pandas==2.0.1
Pillow==9.5.0
protobuf==3.20.3
PyExecJS==1.5.1
PyMySQL==1.0.3
pyparsing==3.0.9
pypiwin32==223
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==306
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.13
sympy==1.12
typing-extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.1.2
wordcloud==1.9.1.1
zipp==3.8.0
